requirements.txt
# ------------------------------
# Core
# ------------------------------
python-dotenv>=1.0.0
redis>=5.0.0
requests>=2.31.0

# ------------------------------
# YouTube / Audio Transcription
# ------------------------------
yt-dlp>=2024.3.10
google-genai>=0.3.0

# ------------------------------
# OCR / PDF Processing
# ------------------------------
pdf2image>=1.17.0
Pillow>=10.0.0
pytesseract>=0.3.10

# ------------------------------
# Google Vertex AI (OCR backend)
# ------------------------------
google-cloud-aiplatform>=1.42.0
vertexai>=1.42.0

google-cloud-storage
pydub>=0.25.1